requirements.txt
# Flask Core
Flask==3.0.0
Werkzeug==3.0.1

# Database
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
cryptography==41.0.7

# Forms & Validation
Flask-WTF==1.2.1
WTForms==3.1.1

# Authentication & Security
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
itsdangerous==2.1.2

# Session Management
Flask-Session==0.5.0

# Excel Export
openpyxl==3.1.2
XlsxWriter==3.1.9

# PDF Generation
reportlab==4.0.7
PyPDF2==3.0.1

# Date/Time Utilities
python-dateutil==2.8.2

# Email (for future notifications)
Flask-Mail==0.9.1

# Environment Variables
python-dotenv==1.0.0

# Development Tools
Flask-DebugToolbar==0.14.1
